Remove extraneous digit from version number
Since the project only uses x.y version numbers, but in the package.json
file it is specified as x.y.z.

This could have been done without using a second variable, but would
basically be using .replace on the same line where version is called for
the embed, which introduces unnecessary overhead (even if small) that
can be eliminated by doing the regex replacement once.