Begin working on #11
While I easily added the code to remove the class roles, it appears that
the messageReactionRemove event still triggers when we remove the user's
reactions, thus causing the already written code to run. While it is
probably faster to just use the few lines I added instead of running the
other function (since we've already fetched a lot of important
information from the API), I can't think of a way to stop this from
happening, and at the very least it allows us to reuse code.