Strange
Copied the example from the Discord.js guide, but it was actually worse
than my implementation. Or at least, it wouldn't always remove the
reactions like it was supposed to? Sticking with this implementation for
now.

The slight change I made, might be better, but it could also be
literally the exact same thing.